moved navbar file to components folder
